+++ gluecodium/src/test/resources/smoke/async/output/dart/lib/smoke.dart
@@ -1,4 +1,5 @@
 export 'src/smoke/async_class.dart' show AsyncClass, AsyncClass$Impl;
 export 'src/smoke/async_renamed.dart' show AsyncRenamed;
 export 'src/smoke/async_struct.dart' show AsyncStruct, AsyncStruct$Impl;
+export 'src/smoke/async_with_skips.dart' show AsyncWithSkips, AsyncWithSkips$Impl;
 export 'src/smoke/throw_me_exception.dart' show ThrowMeException;